fix broken SSR.
